* wx2.9+: fix assert in Compiler::SaveSettings()
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9388 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/compiler.cpp
@@ -516,7 +516,7 @@ void Compiler::SaveSettings(const wxString& baseKey)
         if (i < m_Mirror.RegExes.GetCount() && m_Mirror.RegExes[i] == m_RegExes[i])
             continue;
 
-        group.Printf(_T("%s/regex/re%3.3d"), tmp.c_str(), i + 1);
+        group.Printf(_T("%s/regex/re%3.3lu"), tmp.c_str(), static_cast<unsigned long>(i + 1));
         RegExStruct& rs = m_RegExes[i];
         cfg->Write(group + _T("/description"),  rs.desc,  true);
         if (rs.lt != 0)
